+++ test/t/06_vg_index.t
@@ -7,7 +7,7 @@ PATH=../bin:$PATH # for vg
 
 export LC_ALL="en_US.utf8" # force ekg's favorite sort order 
 
-plan tests 46
+plan tests 47
 
 vg construct -r small/x.fa -v small/x.vcf.gz >x.vg
 
@@ -82,6 +82,8 @@ is $? 0 "building an xg index containing a gPBWT"
 
 xg -i x.xg -x > part.vg
 is "$(cat x.vg part.vg | vg view -j - | jq '.path[].name' | grep '_thread' | wc -l)" 4 "the gPBWT contains the expected number of threads"
+
+is $(vg find -x x.xg -q 1_0 | vg paths -L - | wc -l) 1 "a specific thread may be pulled from the graph by name"
 
 rm -f x.vg x.xg part.vg x.gcsa
